[#175] ♻️refactor : UserSearchCondition 위치 변경
