+++ src/main/java/com/sparta/mypet/domain/auth/UserRepositoryQuery.java
@@ -3,8 +3,8 @@ package com.sparta.mypet.domain.auth;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
+import com.sparta.mypet.domain.auth.dto.UserSearchCondition;
 import com.sparta.mypet.domain.auth.entity.User;
-import com.sparta.mypet.domain.auth.entity.UserSearchCondition;
 
 public interface UserRepositoryQuery {
 	Page<User> findBySearchCond(UserSearchCondition searchCondition, Pageable pageable);
