+++ src/main/java/com/sparta/mypet/domain/auth/UserService.java
@@ -22,12 +22,12 @@ import com.sparta.mypet.domain.auth.dto.SignupRequestDto;
 import com.sparta.mypet.domain.auth.dto.SignupResponseDto;
 import com.sparta.mypet.domain.auth.dto.UserPasswordUpdateRequestDto;
 import com.sparta.mypet.domain.auth.dto.UserResponseDto;
+import com.sparta.mypet.domain.auth.dto.UserSearchCondition;
 import com.sparta.mypet.domain.auth.dto.UserUpdateRequestDto;
 import com.sparta.mypet.domain.auth.dto.UserUpdateResponseDto;
 import com.sparta.mypet.domain.auth.dto.UserWithdrawResponseDto;
 import com.sparta.mypet.domain.auth.entity.User;
 import com.sparta.mypet.domain.auth.entity.UserRole;
-import com.sparta.mypet.domain.auth.entity.UserSearchCondition;
 import com.sparta.mypet.domain.auth.entity.UserStatus;
 import com.sparta.mypet.domain.oauth.SocialAccountService;
 import com.sparta.mypet.domain.oauth.entity.SocialAccount;
